LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/scroll-margin-overflow-ancestor-expected.txt

overflow: visible

overflow: clip

overflow: auto

overflow: auto with scroll padding


PASS scroll-margin is respected when overflowing an overflow: visible container
PASS scroll-margin is respected when overflowing an overflow: clip container
FAIL scroll-margin is respected when overflowing an overflow: auto container assert_equals: gives 40px margin to element slightly above inner scroller expected 960 but got 940
FAIL scroll-margin and scroll-padding are respected when overflowing an overflow: auto container assert_equals: gives 40px margin and 20px padding to element slightly above inner scroller expected 940 but got 920
